test: move test file
